mdlMalloc(): fix size of "size" argument for greater than 2G requests.

+++ null/mdl.c
@@ -371,7 +371,7 @@ void mdlHandler(MDL mdl)
  ** For PVM and most machines these functions are trivial, but on the 
  ** T3D and perhaps some future machines these functions are required.
  */
-void *mdlMalloc(MDL mdl,int iSize)
+void *mdlMalloc(MDL mdl, size_t iSize)
 {	
         if (!iSize) return(NULL);
 	return(malloc(iSize));
